start kind branch
This branch is to add a dynamic dropdown menu to the kind attribute of fiixtures, based on the tutorial at https://simpleisbetterthancomplex.com/tutorial/2018/01/29/how-to-implement-dependent-or-chained-dropdown-list-with-django.html. Currently a little mixed up as the tutorials implementation of class based views is a little differen than what I am familiar with. @TODO: figure out <int:pk> for detail view of individual fixtures.